Another important bugfix.
There was a whitespace problem in the XML parser. A newline and some spaces were 
added to the revision ID number, and since a change to the MediaWiki Flagged 
Revisions extension, this whitespace is a problem. Reviewed articles on dewiki had 
to be reviewed again after a bot made a mistake.

If you have a bot that makes changes on dewiki, please update immediately, as many 
users are complaining about the bug.



git-svn-id: http://svn.wikimedia.org/svnroot/pywikipedia/trunk@5387 6a7f98fc-eeb0-4dc1-a6e2-c2c589a08aa6