Correction for a CPU overload problem introduced with the recent changes in PageGenerators module that would use the DuplicateFilterPageGenerator always (probably a bad idea). This filter was using a 'list' object to check for duplicated 'Page' and was storing 'Page' objects making the comparision process much more complicated...
'set' looks here more appropriate, as it is hashed; and storing for comparision the title and the interwiki link should be enough. This also reduces allocated memory a lot compared with the previous revision (60-65% estimated with a fixed title length of 14 chars).

This commit reduces CPU usage for a so simple task on my five/six years old system from 99% to 30%.

git-svn-id: http://svn.wikimedia.org/svnroot/pywikipedia/trunk@6608 6a7f98fc-eeb0-4dc1-a6e2-c2c589a08aa6